refactor job.strategies to make the file hierarchy more meaningful
